databene-commons-core/src/test/java/org/databene/commons/RoundedNumberFormatTest.java
/*
 * (c) Copyright 2007-2009 by Volker Bergmann. All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without
 * modification, is permitted under the terms of the
 * GNU General Public License.
 *
 * For redistributing this software or a derivative work under a license other
 * than the GPL-compatible Free Software License as defined by the Free
 * Software Foundation or approved by OSI, you must first obtain a commercial
 * license to this software product from Volker Bergmann.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
 * WITHOUT A WARRANTY OF ANY KIND. ALL EXPRESS OR IMPLIED CONDITIONS,
 * REPRESENTATIONS AND WARRANTIES, INCLUDING ANY IMPLIED WARRANTY OF
 * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE OR NON-INFRINGEMENT, ARE
 * HEREBY EXCLUDED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
 * LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
 * CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
 * SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
 * INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
 * CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
 * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 * POSSIBILITY OF SUCH DAMAGE.
 */

package org.databene.commons;

import org.junit.Test;
import static org.junit.Assert.*;

/**
 * Tests the RoundedNumberFormat class.<br/>
 * <br/>
 * Created: 01.09.2007 16:30:06
 * @author Volker Bergmann
 */
public class RoundedNumberFormatTest {

	@Test
    public void test() {
        assertEquals("1,200", RoundedNumberFormat.format(1234, 0));
        assertEquals("1,200", RoundedNumberFormat.format(1234.45, 0));
        assertEquals("1,200", RoundedNumberFormat.format(1234.567, 0));
        assertEquals("1,200.00", RoundedNumberFormat.format(1234.567, 2));
        assertEquals("1,200,000", RoundedNumberFormat.format(1234567, 0));
        assertEquals("1,200,000", RoundedNumberFormat.format(1234567, 2));
    }
	
}
